+++ packages/base-data-service/tests/ExampleDataService.ts
@@ -45,13 +45,26 @@ export type GetActivityResponse = {
   };
 };
 
+export type AddFollowerResponse = {
+  followed: {
+    profileId: string;
+    address: string;
+    name: string;
+    imageUrl?: string | null;
+  }[];
+};
+
 export type PageParam =
   | {
       before: string;
     }
   | { after: string };
 
-const MESSENGER_EXPOSED_METHODS = ['getAssets', 'getActivity'] as const;
+const MESSENGER_EXPOSED_METHODS = [
+  'getAssets',
+  'getActivity',
+  'addFollower',
+] as const;
 
 export class ExampleDataService extends BaseDataService<
   typeof serviceName,
@@ -60,6 +73,8 @@ export class ExampleDataService extends BaseDataService<
   readonly #accountsBaseUrl = 'https://accounts.api.cx.metamask.io';
 
   readonly #tokensBaseUrl = 'https://tokens.api.cx.metamask.io';
+
+  readonly #socialBaseUrl = 'https://social.api.cx.metamask.io';
 
   constructor(
     messenger: ExampleMessenger,
@@ -144,6 +159,30 @@ export class ExampleDataService extends BaseDataService<
       },
       page,
     );
+  }
+
+  async addFollower(followerId: string): Promise<AddFollowerResponse> {
+    return this.executeMutation<AddFollowerResponse>({
+      mutationKey: [`${this.name}:addFollower`, followerId],
+      mutationFn: async () => {
+        const url = new URL(`${this.#socialBaseUrl}/api/v1/users/me/follows`);
+
+        const response = await fetch(url, {
+          method: 'PUT',
+          headers: { 'Content-Type': 'application/json' },
+          body: JSON.stringify({ followerId }),
+        });
+
+        if (!response.ok) {
+          throw new Error(
+            `Mutation failed with status code: ${response.status}.`,
+          );
+        }
+
+        return response.json();
+      },
+      cacheTime: 0, // Not recommended in production, just for testing purposes.
+    });
   }
 
   destroy(): void {